removed venv use, ncnn is no longer avaibale.
