+++ cone_lcm_publisher.py
@@ -16,11 +16,11 @@ if __name__ == '__main__':
     fps = config["fps"]
 
     calibration_data = np.load('cam_calibration_data.npz')
-    ncnn_model = YOLO("../utils/cone_detection_model_ncnn_model")
+    model = YOLO("utils/cone_detection_model.pt")
 
     camera = Camera(camera_id, image_width, image_height, fps)
     register_signal_handlers(camera.cleanup)
-    cone_detector = ConeDetector(ncnn_model, calibration_data)
+    cone_detector = ConeDetector(model, calibration_data)
 
     while camera.running:
         frame = camera.capture_frame()
