--- mbot_vision_lcm/__init__.py
@@ -1,7 +1,0 @@
-"""LCM package __init__.py file
-This file automatically generated by lcm-gen.
-DO NOT MODIFY BY HAND!!!!
-"""
-
-from .mbot_cone_array_t import mbot_cone_array_t
-from .mbot_cone_t import mbot_cone_t
