--- msgs/mbot_cone_t.lcm
@@ -1,8 +1,0 @@
-package mbot_vision_lcm;
-
-struct mbot_cone_t
-{
-    string name;
-    float x;
-    float z;
-}
