+++ tag_cone_detection.py
@@ -7,23 +7,26 @@ from utils.utils import register_signal_handlers
 from utils.config import CAMERA_CONFIG
 from camera.camera import Camera
 from camera.cone_detector import ConeDetector
-
+from camera.apriltag_detector import AprilTagDetector
 """
 Features:
-1. Displays the video live stream with cone detection to the browser.
+1. Displays the video live stream with tag and cone detection to the browser.
 2. Display the pose estimate values.
 
 visit: http://your_mbot_ip:5001
 """
 
-class ConeViewer(Camera):
+class TagConeViewer(Camera):
     def __init__(self, camera_id, width, height, model, calibration_data, fps=None):
         super().__init__(camera_id, width, height, fps)
         self.cone_detector = ConeDetector(model, calibration_data)
+        self.tag_detector = AprilTagDetector(calibration_data)
 
     def process_frame(self, frame):
         self.cone_detector.detect_cones(frame)
         self.cone_detector.draw_cone_detect(frame)
+        self.tag_detector.detect_tags(frame)
+        self.tag_detector.draw_tags(frame)
 
         return frame
 
@@ -43,15 +46,15 @@ if __name__ == '__main__':
     calibration_data = np.load('cam_calibration_data.npz')
 
     # Load a YOLO11n PyTorch model
-    # model = YOLO("../utils/cone_detection_model.pt")
+    model = YOLO("utils/cone_detection_model.pt")
 
     # Export the model to NCNN format
     # model.export(format="ncnn")
 
     # Load the exported NCNN model
-    ncnn_model = YOLO("../utils/cone_detection_model_ncnn_model")
+    # ncnn_model = YOLO("utils/cone_detection_model_ncnn_model")
 
-    camera = ConeViewer(camera_id, image_width, image_height,
-                                    ncnn_model, calibration_data, fps)
+    camera = TagConeViewer(camera_id, image_width, image_height,
+                                    model, calibration_data, fps)
     register_signal_handlers(camera.cleanup)
     app.run(host='0.0.0.0', port=5001)
